processing-block-model extracted to own file
