+++ common/CMakeLists.txt
@@ -47,6 +47,8 @@ set(COMMON_SRC
     "${CMAKE_CURRENT_LIST_DIR}/option-model.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/subdevice-model.h"
     "${CMAKE_CURRENT_LIST_DIR}/subdevice-model.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/processing-block-model.h"
+    "${CMAKE_CURRENT_LIST_DIR}/processing-block-model.cpp"
     )
 
 set(SW_UPDATE_FILES
